+++ packages/dev-dependencies/patches/index.js
@@ -1,2 +1,5 @@
 exports.majorBrickNext = require("./majorBrickNext");
 exports.updateLernaAllowBranch = require("./updateLernaAllowBranch");
+exports.updateMRTemplates = require("./updateMRTemplates");
+exports.updateBuildStories = require("./updateBuildStories");
+exports.updateRenovateFileFilters = require("./updateRenovateFileFilters");