+++ packages/dev-dependencies/patches/updateMRTemplates.spec.js
@@ -1,0 +1,14 @@
+const updateMRTemplates = require("./updateMRTemplates");
+const fs = require("fs-extra");
+
+jest.spyOn(fs, "outputFileSync").mockImplementation(() => void 0);
+
+describe("updateMRTemplates", () => {
+  it("should work", () => {
+    updateMRTemplates();
+    expect(fs.outputFileSync).toBeCalledWith(
+      expect.any(String),
+      expect.stringContaining("依赖检查")
+    );
+  });
+});