+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.6](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.5...@easyops/build-config-factory@2.0.6) (2020-09-28)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [2.0.5](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.4...@easyops/build-config-factory@2.0.5) (2020-09-28)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [2.0.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.3...@easyops/build-config-factory@2.0.4) (2020-09-27)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [2.0.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.2...@easyops/build-config-factory@2.0.3) (2020-09-25)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
 ## [2.0.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.1...@easyops/build-config-factory@2.0.2) (2020-09-17)
 
 **Note:** Version bump only for package @easyops/build-config-factory