+++ packages/create-next-repo/CHANGELOG.md
@@ -3,6 +3,69 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.22](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.21...@easyops/create-next-repo@1.0.22) (2020-09-28)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.21](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.20...@easyops/create-next-repo@1.0.21) (2020-09-28)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.20](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.19...@easyops/create-next-repo@1.0.20) (2020-09-28)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.19](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.18...@easyops/create-next-repo@1.0.19) (2020-09-28)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.18](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.17...@easyops/create-next-repo@1.0.18) (2020-09-28)
+
+### Bug Fixes
+
+- update renovate file filters ([1310c23](https://git.easyops.local/anyclouds/next-core/commits/1310c23))
+
+## [1.0.17](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.16...@easyops/create-next-repo@1.0.17) (2020-09-28)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.16](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.15...@easyops/create-next-repo@1.0.16) (2020-09-28)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.15](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.14...@easyops/create-next-repo@1.0.15) (2020-09-27)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.14](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.13...@easyops/create-next-repo@1.0.14) (2020-09-27)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.13](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.12...@easyops/create-next-repo@1.0.13) (2020-09-27)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.12](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.11...@easyops/create-next-repo@1.0.12) (2020-09-27)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.11](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.10...@easyops/create-next-repo@1.0.11) (2020-09-27)
+
+### Bug Fixes
+
+- update MR templates ([95e2281](https://git.easyops.local/anyclouds/next-core/commits/95e2281))
+
+## [1.0.10](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.9...@easyops/create-next-repo@1.0.10) (2020-09-25)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [1.0.9](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.8...@easyops/create-next-repo@1.0.9) (2020-09-25)
+
+### Bug Fixes
+
+- add legacy branches into lerna allow branches ([4fdf2c8](https://git.easyops.local/anyclouds/next-core/commits/4fdf2c8))
+- upgrade template dependencies ([b4dce22](https://git.easyops.local/anyclouds/next-core/commits/b4dce22))
+
 ## [1.0.8](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.7...@easyops/create-next-repo@1.0.8) (2020-09-18)
 
 **Note:** Version bump only for package @easyops/create-next-repo