+++ packages/dll-differ/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dll-differ@1.0.0...@easyops/dll-differ@1.0.1) (2020-09-25)
+
+**Note:** Version bump only for package @easyops/dll-differ
+
 # [1.0.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dll-differ@1.0.0-next.3...@easyops/dll-differ@1.0.0) (2020-09-13)
 
 **Note:** Version bump only for package @easyops/dll-differ